move regester to AuthenticationController
